[escher/stack_view] Increase the StackView max depth limit to 16.

+++ escher/include/escher/stack_view.h
@@ -11,7 +11,7 @@ class StackView : public View {
  public:
   enum class Style { GrayGradation, PurpleWhite, WhiteUniform };
 
-  using Mask = uint8_t;
+  using Mask = uint16_t;
 
   static constexpr uint8_t k_maxDepth = sizeof(Mask) * 8;
